Pull out app and mysql generation to lib files
